Remove unused code and fix tests
