+++ plugins/user-authenticators/oauth2/src/main/java/org/apache/cloudstack/oauth2/api/response/OauthProviderResponse.java
@@ -1,17 +1,62 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
 package org.apache.cloudstack.oauth2.api.response;
 
 import com.cloud.serializer.Param;
 import com.google.gson.annotations.SerializedName;
-import org.apache.cloudstack.api.response.AuthenticationCmdResponse;
+import org.apache.cloudstack.api.ApiConstants;
+import org.apache.cloudstack.api.BaseResponse;
+import org.apache.cloudstack.api.EntityReference;
+import org.apache.cloudstack.oauth2.vo.OauthProviderVO;
 
-public class OauthProviderResponse extends AuthenticationCmdResponse {
-    @SerializedName("id")
-    @Param(description = "The Oauth Provider ID")
+@EntityReference(value = OauthProviderVO.class)
+public class OauthProviderResponse extends BaseResponse {
+
+    @SerializedName(ApiConstants.ID)
+    @Param(description = "ID of the provider")
     private String id;
 
-    @SerializedName("providerName")
-    @Param(description = "Oauth Provider name")
-    private String providerName;
+    @SerializedName(ApiConstants.PROVIDER)
+    @Param(description = "Name of the provider")
+    private String provider;
+
+    @SerializedName(ApiConstants.DESCRIPTION)
+    @Param(description = "Description of the provider registered")
+    private String description;
+
+    @SerializedName(ApiConstants.CLIENT_ID)
+    @Param(description = "Client ID registered in the OAuth provider")
+    private String clientId;
+
+    @SerializedName(ApiConstants.REDIRECT_URI)
+    @Param(description = "Redirect URI registered in the OAuth provider")
+    private String redirectUri;
+
+    @SerializedName(ApiConstants.ENABLED)
+    @Param(description = "Whether the OAuth provider is enabled or not")
+    private boolean enabled;
+
+    public OauthProviderResponse(String id, String provider, String description, String clientId, String redirectUri) {
+        this.id = id;
+        this.provider = provider;
+        this.description = description;
+        this.clientId = clientId;
+        this.redirectUri =  redirectUri;
+    }
 
     public String getId() {
         return id;
@@ -21,11 +66,44 @@ public class OauthProviderResponse extends AuthenticationCmdResponse {
         this.id = id;
     }
 
-    public String getProviderName() {
-        return providerName;
+    public String getProvider() {
+        return provider;
     }
 
-    public void setProviderName(String providerName) {
-        this.providerName = providerName;
+    public void setProvider(String provider) {
+        this.provider = provider;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getClientId() {
+        return clientId;
+    }
+
+    public void setClientId(String clientId) {
+        this.clientId = clientId;
+    }
+
+    public boolean getEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    public String getRedirectUri() {
+        return redirectUri;
+    }
+
+    public void setRedirectUri(String redirectUri) {
+        this.redirectUri = redirectUri;
     }
 }
